[OC-1538] Moved all logfile moveing logic into lifecycle management handler
