+++ src/backend/src/main/java/com/becon/opencelium/backend/execution/supportfile/SupportFileService.java
@@ -8,6 +8,5 @@ public interface SupportFileService {
     List<SupportFile> connectionSupportFileList(Long connectionId);
     File getSupportFile(Long connectionId, String zipFileName);
     File getSupportFile(Long connectionId);
-    void collectFiles(Long connectionId, long executionId, String timestamp, String type);
     void deleteSupportFile(String zipFileName);
 }
